requirements.txt
# requests is library to make the HTTP requests
requests>=2.18.4

# libusb1 monitors USB devices
libusb1>=1.6.4

# sh simplifies shell commands 
sh>=1.12.14

# For the path monitoring
watchdog
